requirements.txt
# Image processing
Pillow>=9.0
opencv-python>=4.5
numpy<2.0.0
scikit-image>=0.19.0

# Environment variables
# python-dotenv==1.0.1

# Type checking
mypy==1.9.0
types-Pillow==10.2.0.20240324

# Testing
pytest==8.0.2
pytest-cov==4.1.0

# Linting
black==24.2.0
isort==5.13.2
flake8==7.0.0

# 開発ツール
pip-licenses==4.3.3  # ライセンス管理用
